Prevent sending email with duplicate "To" headers (#1295)
Currently the "To" header for sent emails is sent to the `mail` function both in the `to` argument and manually in the `additional_headers` argument. This results in multiple "To" headers in the resulting email.

Various email providers (e.g. GMail or Hotmail/Outlook/Live.com) deal with this duplicate header in different ways.

GMail for example sees that there are duplicate "To" headers and collapses them so you end up with a "To" header of "John Doe <jdoe@example.com>"

Hotmail/Outlook/Live.com combines them so you end up with a "To" header of "<jdoe@example.com>, John Doe <jdoe@example.com>"

This comes into play and matters in regards to DKIM signatures in emails. With these differing handling of this invalid set of email headers the result is that DKIM signed emails show as having valid signatures in GMail and invalid signatures in Hotmail/Outlook/Live.com

Email sent to Hotmail/Outlook/Live.com is consequently marked as junk/spam because it has an invalid DKIM signature

This commit changes how the php `mail` function is called by
* removing the manually set "To" header
* passing in the full "First Last <flast@example.com>" RFC 2822 compliant string as the `to` argument

With this, DKIM signatures show as valid.

This change doesn't however do anything about validating or encoding the string pass as the `to` argument which may need further work.

Co-authored-by: Geoff Humphrey <geoffhumphrey@users.noreply.github.com>