Address Issue #1519; MHP Organizer Report Bug Fix
Correct reported error with MHP export incorrectly attributing someentries/scores/places to entrants listed in the report. Should not be a factor for most installations; only those whose ids in the user table are not identical to the uids in the brewer table. Also changed the updated date method for user accounts and entries to use an SQL-friendly format of the current UNIX timestamp, instead of the SQL NOW() function that may record the time code of the server running the instance. Minor, but updated dates/times were not correct.